refactor(#3269): extract for-await-helpers.ts sync-drive helpers from loops.ts (#3182)
Verbatim move of FOR_AWAIT_SYNC_DRIVE_STEP_CAP + emitForAwaitElementUnwrap +
emitForAwaitStepCapCheck. Trimmed 5 now-unused imports.
emit-identity IDENTICAL, tsc 0.

✓
